:art: [#840] -- move signal handlers to separate module
